proforma-generator/requirements.txt
fpdf2==2.7.6
Pillow==10.1.0

